requirements.txt
Flask
python-dotenv
requests
PyPDF2
python-docx
firebase-admin
openai
google-generativeai